+++ src/main/java/it/gov/pagopa/nodetsworker/service/WorkerService.java
@@ -4,9 +4,9 @@ import it.gov.pagopa.nodetsworker.models.*;
 import it.gov.pagopa.nodetsworker.repository.CosmosBizEventClient;
 import it.gov.pagopa.nodetsworker.repository.CosmosNegBizEventClient;
 import it.gov.pagopa.nodetsworker.repository.CosmosVerifyKOEventClient;
-import it.gov.pagopa.nodetsworker.repository.model.NegativeBizEvent;
-import it.gov.pagopa.nodetsworker.repository.model.PositiveBizEvent;
-import it.gov.pagopa.nodetsworker.repository.model.VerifyKOEvent;
+import it.gov.pagopa.nodetsworker.repository.models.NegativeBizEvent;
+import it.gov.pagopa.nodetsworker.repository.models.PositiveBizEvent;
+import it.gov.pagopa.nodetsworker.repository.models.VerifyKOEvent;
 import it.gov.pagopa.nodetsworker.resources.response.TransactionResponse;
 import it.gov.pagopa.nodetsworker.util.ValidationUtil;
 import jakarta.enterprise.context.ApplicationScoped;
